fix: check perms before write, only write if allowed
It is not enough to just check vim's `'ro'` option and only write a
buffer when it is unset. For one, `'ro'` may be (or become) unset even
when the file is unwritable. By default, vim also unsets `'ro'` after
writing a readonly buffer, regardless of whether the file permissions
have changed or not (see also the `Z` flag from the `'cpo'` option).
Even when unset, vim will prompt (or error if `'confirm'` is unset)
every time to write a buffer whose file has insufficient permissions to
do a simple write (i.e. a write w/o resorting to a `cp` or `mv` of the
file), which can quickly become annoying when the writing is done
frequently/automatically.

Additional Comments

In the future, consider allowing the user to opt-into forcefully writing
readonly buffers. There could also be an option to ignore any resulting
errors in this case as well. These options might not be very safe
though, and they would require `W` to be absent from `'cpo'`.