perf: use `getbufinfo()` over `nvim_list_bufs()`
The former is faster at filtering buffers than using the latter while
filtering manually (in Lua). The former is faster both cold (first
call), and after warmup, and overall is at least 10% quicker in this
situation.

This change improves the performance of the timer callback which goes
through all buffers and saves those that are modified and meet the other
criteria.